+++ lib/sequence/client.rb
@@ -6,6 +6,7 @@ require_relative './asset'
 require_relative './balance'
 require_relative './contract'
 require_relative './dev_utils'
+require_relative './feed'
 require_relative './flavor'
 require_relative './key'
 require_relative './stats'
@@ -87,6 +88,11 @@ module Sequence
     # @return [Transaction::ClientModule]
     def transactions
       @transactions ||= Transaction::ClientModule.new(self)
+    end
+
+    # @return [Feed::ClientModule]
+    def feeds
+      @feeds ||= Feed::ClientModule.new(self)
     end
 
     # @return [Contract::ClientModule]